Some fixes here and there on variables. Added MyBlue and MyRed
Added MyBlue and MyRed to keep them as options for setting the configuration names "redtile" and "bluetile" (physical prototypes worked on). Also added a public variable to MyTile called ConfigNameToSet that can be set in the inspector to set the config name if needed if device name isn't used. A number of leds public variable was added as well if it's needed.