Documentation, Event names, RFID dummy tests, folder structure changes
Adding documentation for installation and usage.

Changed event name calls from the device components, and the event listener calls on all classes. 

Created an RFID dummy test where an Arduino sends a set string every 30th second, which is added to the RFID last read variable on the RFID device component in Unity. 

Will change folder structure to remove the Unity Project files and only keep the folder with the framework files. This should make importing or opening in newer versions of Unity easier.